requirements.txt
streamlit #frontend
google-generativeai  #GEnAi model
python-dotenv #meant for local environment